Data.OrdMonad

Description

A simple implementation of Functor and Monad with Ord constraints, to make operations on those data types more efficient.

Functors and Monads with Ordering

Functors with ordering

class OrdFunctor (f :: Type -> Type)

Functors with an additional Ord constraint. The primary use case for the OrdFunctor is to treat data structures like Set as a functor, where the Ord-constraint is used for performance reasons. Implementations of ordMap should adhere to the same laws as for fmap for a Functor F:

Identity

ordMap id == id

Composition

ordMap (f . g) == ordMap f . ordMap g

The composition law is only required if the extensionality property of type class Eq holds for the domain of f. Effectively, this states that Eq should behave like proper equality for f, or conversely, if x == y and f x /= f y, then compositionality is not expected to hold.

Minimal complete definition

ordMap

ordMap :: (OrdFunctor f, Ord b) => (a -> b) -> f a -> f b

Map a function over a functorial type, just like fmap, but with an additional Ord constraint.

(<#>) :: (OrdFunctor f, Ord b) => (a -> b) -> f a -> f b

An infix synonym for ordMap, similar to <$>.

Monads with ordering

class OrdFunctor m => OrdMonad (m :: Type -> Type)

Monads with an additional Ord constraint. Analogously to how an OrdFunctor specializes a regular Functor, OrdMonad uses the Ord-constraint is used for performance reasons. Any instance should adhere to the Monad laws, assuming the extensionality property for equality Eq. See OrdFunctor for details.

Minimal complete definition

ordReturn, ordBind

ordBind :: (OrdMonad m, Ord b) => m a -> (a -> m b) -> m b

Return operation, similar to the standard monadic return. No Ord constraint is present here, as comparing values is not needed for injecting a single value into a monadic type.

ordReturn :: OrdMonad m => a -> m a

Bind operation, using an Ord constraint, similar to the standard monadic bind operation >>=.

ordJoin :: (Ord a, OrdMonad m) => m (m a) -> m a

Standard monadic join, but with an additional Ord constraint.