Lattest.Adapter.Adapter

Definition

data Adapter act i

An adapter to a (usually external) system. Uses two channels for interaction: one to send input commands, and one to receive outputs.

Constructors

Adapter

Fields

inputCommandsToSut :: OutputStream i

Channel for sending input commands.

actionsFromSut :: TInputStream act

Channel for receiving outputs.

close :: IO ()

Close the (channels of this) adapter.

Interaction

send :: i -> Adapter act i -> IO ()

Send an input to the adapter.

observe :: Adapter act i -> IO (Maybe act)

Observe in a blocking manner, with two possible outcomes:

An observation is made,

The stream has closed, and any further attempts will give the same result.

tryObserve :: Adapter act i -> IO (Streamed act)

Try to observe, without blocking, with three possible outcomes:

An observation is made,

no observation was ready, so no observation was made but another attempt may make an observation,

The stream has closed, and any further attempts will give the same result.

Transformations

map :: (i' -> i) -> (act -> act') -> Adapter act i -> IO (Adapter act' i')

Map function over both the inputs sent to, and the outputs received from, the adapter.

mapInputCommands :: (i' -> i) -> Adapter act i -> IO (Adapter act i')

Map a function over the inputs sent to the adapter.

mapActionsFromSut :: (act -> act') -> Adapter act i -> IO (Adapter act' i)

Map a function over the outputs received from the adapter.

parseActionsFromSut :: Parser act -> Adapter ByteString i -> IO (Adapter act i)

Transform a raw adapter which receives ByteStrings to an adapter which receives objects, parsing the bytestrings with the given parser.