Lattest.Adapter.StandardAdapters

Description

This module contains building blocks for constructing out-of-the-box Adapters.

Note that adapters do not observe inputs by default. In other words, an input command from a test controller does not automatically lead to an input transition in the specification model. Use acceptingInputs to observe all sent inputs, or acceptingInputsWithIncompletenessAsFailures to also observe failures in processing inputs in the adapter. The latter is useful if the processing of inputs is defined partially, in which case some inputs cannot be sent to the adapter.

Type Definition

data Adapter act i

An adapter to a (usually external) system. Uses two channels for interaction: one to send input commands, and one to receive outputs.

Ready-to-use Adapters

Pure adapters

Adapters which are defined in haskell itself.

pureMealyAdapter :: (state -> i -> state) -> (state -> i -> [act]) -> state -> IO (Adapter act i)

adapter which, after every input, directly sends the corresponding sequence of outputs

pureAdapter :: (Ord i, Ord o, RandomGen g) => g -> Double -> (state -> Map (IOAct i o) state) -> state -> IO (Adapter (TimeoutIF i o) (Maybe i))

Adapter which, after every action, has the given probability of producing non-deterministically one of the corresponding (non-timeout) output transitions if any is available. After receiving a Nothing input, an output will be produced, which may be a timeout.

Socket Adapters

Settings

data SocketSettings act i

Settings for a socket Adapter.

Constructors

SocketSettings

Fields

hostName :: HostName

portNumber :: PortNumber

baseSocketSettings :: SocketSettings act i

Default settings for a socket Adapter.

Base Socket Adapters

connectSocketAdapter :: IO (Adapter ByteString ByteString)

Create an adapter by connecting to a server socket, with the default settings.

connectSocketAdapterWith :: SocketSettings act i -> IO (Adapter ByteString ByteString)

Create an adapter by connecting to a server socket, with the given settings.

Socket Adapters with JSON

connectJSONSocketAdapter :: (ToJSON i, FromJSON o) => IO (Adapter o i)

Create an adapter by connecting to a server socket, with the default settings, and sending inputs and reading outputs in JSON format.

connectJSONSocketAdapterWith :: (ToJSON i, FromJSON o) => SocketSettings act i -> IO (Adapter o i)

Create an adapter by connecting to a server socket, with the given settings, and sending inputs and reading outputs in JSON format.

connectJSONSocketAdapterAcceptingInputs :: (ToJSON i, FromJSON o) => IO (Adapter (IOAct i o) i)

Create an adapter by connecting to a server socket, with the default settings, and sending inputs and reading outputs in JSON format, observing any input as accepted.

connectJSONSocketAdapterAcceptingInputsWith :: (ToJSON i, FromJSON o) => SocketSettings act i -> IO (Adapter (IOAct i o) i)

Create an adapter by connecting to a server socket, with the given settings, and sending inputs and reading outputs in JSON format, observing any input as accepted.

Transformations on Adapters

Encoding and Decoding

encodeUtf8 :: Adapter act ByteString -> IO (Adapter act String)

Take an adapter that sends raw ByteStrings, and transform it to an adapter that sends Strings encoded in utf-8.

decodeUtf8 :: Adapter ByteString i -> IO (Adapter String i)

Take an adapter that receives raw ByteStrings, and transform it to an adapter that receives Strings decoded from utf-8.

endecodeUtf8 :: Adapter ByteString ByteString -> IO (Adapter String String)

Take an adapter that sends and receives raw ByteStrings, and transform it to an adapter that sends and receives Strings, encoded in utf-8 and decoded from utf-8.

encodeJSONInputCommands :: ToJSON i => Adapter act ByteString -> IO (Adapter act i)

Take an adapter that sends raw ByteStrings, and transform it to an adapter that sends any type encoded in JSON.

parseJSONActionsFromSut :: FromJSON act => Adapter ByteString i -> IO (Adapter act i)

Take an adapter that receives raw ByteStrings, and transform it to an adapter that receives any type decoded from JSON.

Observing Inputs

acceptingInputs :: Adapter o i -> IO (Adapter (IOAct i o) i)

Transform a given Adapter to accept all inputs. In other words, every selected input command directly becomes an observed input action. Every observed output from the given Adapter becomes an output action.

acceptingInputsWithIncompletenessAsFailures :: Adapter o i -> IO (Adapter (IFAct i o) i)

Transform a given Adapter to accept all inputs, but if the given Adapter would call an undefined case in a partial function, then interpret this as an observed input failure. Any input command that does not call undefined cases of partial functions directly become observed input action. Every observed output from the given Adapter becomes an output action.

Input failures are useful for Adapters written in pure haskell. Instead of only writing complete functions to process input commands, where unexpected input commands may e.g. lead to an error output, the adapter may also just process input commands with partial functions instead, avoiding some boilerplate code.

Observing Timeouts

withTimeout :: NominalDiffTime -> Adapter (IOAct i o) i -> IO (Adapter (TimeoutIO i o) (Maybe i))

Transform the given Adapter by introducing timeout observations. A timeout is observed after the given timeout duration, after any other observation.

withTimeoutMillis :: Int -> Adapter (IOAct i o) i -> IO (Adapter (TimeoutIO i o) (Maybe i))

Transform the given Adapter by introducing timeout observations. A timeout is observed after the given number of milliseconds, after any other observation.