Lattest.Exec.StandardTestControllers

Description

This module contains building blocks for constructing out-of-the-box TestControllers.

TestControllers have multiple duties during testing:

selecting test inputs,

deciding whether to continue testing,

returning testing results (other than Pass or Fail), and

potentially performing side effects during testing.

The building blocks in this module allow composing TestControllers in a modular way, by combining various choices for these responsibilities. Every building block carries its own state. For example, to stop testing after a fixed number of steps, a state in the form of a counter is needed, whereas returning the observed trace as test result requires recording the observed trace as state.

Test Selectors

selector :: InputCommand i act => state -> (state -> AutSem m loc q t tloc act -> m q -> IO (Maybe (i, state))) -> (state -> AutSem m loc q t tloc act -> act -> m q -> IO (Maybe state)) -> TestSelector m loc q t tloc act state i

Create a TestSelector. Requires one function to select an input, also updating the state, and one function to update the state when observing an action.

type TestSelector m loc q t tloc act s i = TestController m loc q t tloc act s i ()

Testselectors are test controllers that are only concerned with selecting inputs for testing. They do not return any testing results.

randomTestSelector :: (AutomatonSemantics m loc q t tloc act, FiniteMenu t act, Foldable m, InputCommand i act) => IO (TestSelector m loc q t tloc act StdGen i)

A TestSelector that picks inputs uniformly pseudo-randomly from the outgoing transitions from the current state configuration.

randomTestSelectorFromSeed :: (AutomatonSemantics m loc q t tloc act, FiniteMenu t act, Foldable m, InputCommand i act) => Int -> TestSelector m loc q t tloc act StdGen i

A TestSelector that picks inputs uniformly pseudo-randomly from the outgoing transitions from the current state configuration, starting with the given random seed.

randomTestSelectorFromGen :: (AutomatonSemantics m loc q t tloc act, FiniteMenu t act, Foldable m, InputCommand i act, RandomGen g) => g -> TestSelector m loc q t tloc act g i

A TestSelector that picks inputs uniformly pseudo-randomly from the outgoing transitions from the current state configuration, based on the given random generator.

Stop Conditions

type StopCondition m loc q t tloc act s = TestController m loc q t tloc act s () ()

StopConditions are test controllers that are only concerned with deciding whether to continue testing after observing an action. They do not select inputs or return any testing results.

stopCondition :: s -> (s -> AutSem m loc q t tloc act -> act -> m q -> IO (Maybe s)) -> StopCondition m loc q t tloc act s

Create a state-based stop condition, starting in the given initial state. The provided function should provide either Just a new state to continue testing, or Nothing to stop testing.

untilCondition :: TestController m loc q t tloc act state1 i1 r1 -> TestController m loc q t tloc act state2 i2 r2 -> TestController m loc q t tloc act (state1, state2) i1 r1

Apply a stop condition: run the (first) test controller until it returns a result, or until the stop condition is reached. The selected inputs and returned result come from the test controller.

stopAfterSteps :: Int -> StopCondition m loc q t tloc act Int

Observe a fixed number of actions, and then stop testing.

Note: since stop conditions only decide whether to stop testing after observing an action, the minimum number of actions to observe is 1.

Test Observers

type TestObserver m loc q t tloc act s r = TestController m loc q t tloc act s () r

TestObservers are only concerned with returning a result after testing. They do not select inputs or decide whether to continue testing.

observer :: s -> (s -> AutSem m loc q t tloc act -> act -> m q -> IO s) -> (s -> IO r) -> TestObserver m loc q t tloc act s r

Create a TestObserver.

andObservingWith :: TestController m loc q t tloc act state1 i1 r1 -> (r1 -> r2 -> r) -> TestController m loc q t tloc act state2 i2 r2 -> TestController m loc q t tloc act (state1, state2) i1 r

Apply an observer: Use the (former) test controller, but also returning the result of the (latter) observer. Combine the two results using the given function.

andObserving :: TestController m loc q t tloc act state1 i1 r1 -> TestController m loc q t tloc act state2 i2 r2 -> TestController m loc q t tloc act (state1, state2) i1 (r1, r2)

Apply an observer: use the (former) test controller, but also returning the result of the (latter) observer in a tuple.

observingOnly :: TestController m loc q t tloc act state1 i1 r1 -> TestController m loc q t tloc act state2 i2 r2 -> TestController m loc q t tloc act (state1, state2) i1 r2

Apply an observer: use the (former) test controller, but only return the result of the (latter) observer.

traceObserver :: TestObserver m loc q t tloc act [act] [act]

A TestObserver that returns the trace of all observed actions

stateObserver :: TestObserver m loc q t tloc act (Maybe (m q)) (Maybe (m q))

A TestObserver that returns the last state configuration of the specification model.

Test Side Effects

type TestSideEffect m loc q t tloc act s = TestController m loc q t tloc act s () ()

TestSideEffects perform side effects during testing, but have no impact on the testing itself, nor on the result.

withSideEffect :: TestController m loc q t tloc act state1 i1 r1 -> TestController m loc q t tloc act state2 i2 r2 -> TestController m loc q t tloc act (state1, state2) i1 r1

Apply a side effect: Use the (former) test controller, but also perform the side effect of the (latter) observer.

testSideEffect :: s -> (s -> AutSem m loc q t tloc act -> act -> m q -> IO s) -> TestSideEffect m loc q t tloc act s

Create a TestSideEffect. The provided function returns the new state in an IO monad that can also perform the side effects.

printActions :: Show act => TestSideEffect m loc q t tloc act ()

Print observed to stdin actions during testing.

printState :: Show (m q) => TestSideEffect m loc q t tloc act ()

Print the state configuration of the specification during testing.