Lattest.Exec.StandardTestControllers.CompleteTestSuite

Documentation

accessSeqSelector :: (Ord q, Eq i, Eq o, Show i, Show o) => ConcreteSuspAutIntrpr Det q i o -> q -> TestController Det q q (IOAct i o) () (IOSuspAct i o) [IOAct i o] (Maybe i) Bool

A TestController that selects inputs that lead to the given targetState. If unexpected outputs are selected by the SUT the TestSelector still tries to provide the inputs of the access sequence, but this may result in reaching another state. Result Bool is True when access sequence has been followed and false when the SUT deviated

adgTestSelector :: (Ord q, Ord l, Eq l, Show l) => ConcreteSuspAutIntrpr Det q l l -> l -> TestController Det q q (IOAct l l) () (IOSuspAct l l) (Evidence l) (Maybe l) (Set q)

A TestController that selects inputs according to the adaptive distinguishing sequence of the given automaton

nCompleteSingleState :: (Ord q, Ord l, Show l) => ConcreteSuspAutIntrpr Det q l l -> Int -> Int -> l -> q -> TestController Det q q (IOAct l l) () (IOSuspAct l l) (((), [IOSuspAct l l]), Maybe (Det q)) i20 ([IOSuspAct l l], Maybe (Det q)) -> IO (TestController Det q q (IOAct l l) () (IOSuspAct l l) (Either (Either [IOAct l l] StdGen, Int) (Evidence l), (((), [IOSuspAct l l]), Maybe (Det q))) (Maybe l) ([IOSuspAct l l], Maybe (Det q)))

A TestController that yields tests that tries to take an access sequence to the targetState and then executes the adaptive distinguishing sequence

runNCompleteTestSuite :: (Ord q, Ord l, Show q, Show l) => IO (Adapter (IOAct l l) l) -> AutSyntax Det q (IOAct l l) () -> Int -> l -> [(q, Int)] -> IO [(q, Verdict, ([IOSuspAct l l], Maybe (Det q)))]

Runs tests from nCompleteSingleState for each given targetState and seed