Lattest.Exec.Testing

Description

This module contains the main functions and data structures to run experiments against (external) systems, specifically testing experiments.

For an experiment, we assume that we have an Adapter, which serves as an interface to send inputs to a system under test, and observe actions shown by that system. The experiment is controlled by an ActionController. Specifically for a testing experiment, we also assume that we have a specification model that dictates which observations are allowed and forbidden, and a test controller that steers the testing.

To get started with testing experiments quickly, use the runTester function, and to get the required ingredients, see the following modules:

Lattest.Adapter.StandardAdapters to create adapters,

Lattest.Exec.StandardTestControllers to create test controllers, and

Lattest.Model.StandardAutomata to create specification models.

Generic Experiments

Can be used to let an action controller communicate with a system under test, resulting in some arbitrary result.

data ActionController act i r state

The controller of an experiment.

Constructors

ActionController

Fields

controllerState :: state

Any state that the controller needs for its decision making duties.

select :: state -> IO (Either (i, state) r)

Select an input and compute the new state, or decide to stop the experiment and return a result.

update :: state -> act -> IO (Either state r)

Provided an observed action, compute the new state or stop the experiment and return a result.

handleClose :: state -> IO r

Handle the end of the action stream, i.e. the other end closing, ending the experiment.

runExperiment :: ActionController act i r state -> Adapter act i -> IO r

Run an experiment by interacting with the given adapter, controlled by the given action controller. The experiment stops when the action controller decides to. Returns the result returned by the action controller.

Testing Experiments

Specific form of experiments, in which an automaton specification model describes the intended behaviour that the system under test should adhere to, and the test controller sends inputs to observe whether the system under test actually conforms to that specification model.

This is also known as Model based testing, see e.g.

Jan Tretmans, Model based testing with labelled transition systems (Formal Methods and Testing), 2008

data TestController (m :: Type -> Type) loc q t tloc act state i r

The controller of a testing experiment. The tester may return a result at the end of a testing experiment. Note that it does not need to return a verdict Pass or Fail, this verdict is automatically inferred based on the specification model. Results can be used to record additional information that the tester is interested in, depending on the controller implementation. For example, the actions observed during a testing experiment, or whether certain coverage criteria were achieved during the experiment.

Constructors

TestController

Fields

testControllerState :: state

Any state that the controller needs for its decision making duties.

selectTest :: InputCommand i act => state -> AutSem m loc q t tloc act -> m q -> IO (Either (i, state) r)

Select a test based on test controller state, the specification (in its current state), and previous specification configuration. Either select a new controller state and a input, or stop testing and return a result from the controller.

updateTestController :: state -> AutSem m loc q t tloc act -> act -> m q -> IO (Either state r)

Select a test based on test controller state, the specification (in its current state), an observed action, and previous specification configuration. Either select a new controller state, or stop testing and return a result from the controller.

handleTestClose :: state -> IO r

Handle the end of the action stream, i.e. the other end closing, ending the experiment.

makeTester :: forall (m :: Type -> Type) loc q t tloc act i state r. (AutomatonSemantics m loc q t tloc act, InputCommand i act, PermissionConfiguration m) => AutSem m loc q t tloc act -> TestController m loc q t tloc act state i r -> ActionController act i (Verdict, r) (AutSem m loc q t tloc act, TestController m loc q t tloc act state i r)

From a test controller and a specification model, create an action controller. The test controller is used to decide which inputs are supplied to the system under test, and whether to continue or stop testing. The automaton specification model is used to infer whether observed actions are allowed or not, and to return a verdict in case of forbidden or underspecified observations.

runTester :: forall (m :: Type -> Type) loc q t tloc act i state r. (AutomatonSemantics m loc q t tloc act, InputCommand i act, PermissionConfiguration m) => AutSem m loc q t tloc act -> TestController m loc q t tloc act state i r -> Adapter act i -> IO (Verdict, r)

Running a tester requires:

a specification model in the form of an automaton, as defined in Lattest.Model.Automaton

a TestController, defined in this module itself, and

an adapter, as defined in Lattest.Adapter.Adapter.

Running a testing experiment is done by interacting with the given adapter, controlled by the given test controller. The experiment stops when the action controller decides to, or when an observation is made that is forbidden or underspecified according to the specification model. Returns the test verdict according to the specification model and the additional result returned by the test controller.

data Verdict

The verdict resulting from a testing experiment: did the system under test pass the test?

Constructors

Pass

Fail