Lattest.Model.Alphabet

Description

This module contains the definitions and semantics of different forms of observable actions, and inputs used in testing experiments.

Translating between Actions and Inputs

class TestChoice i act

If an input type is an TestChoice to a type of observable actions, this means that

given such an input, it is possible to derive the corresponding observable action, or sequence of actions, and

an observable action may (but also may not) correspond to an input.

Minimal complete definition

actToChoice, choiceToActs

choiceToActs :: TestChoice i act => i -> [act]

Derive the sequence of observable actions that correspond to an input.

actToChoice :: TestChoice i act => act -> Maybe i

If an observable action corresponds to an input, then derive that input.

Action types

Inputs and Outputs

data IOAct i o

Observable actions that may be either inputs provided to a system, or outputs from that system.

Constructors

In i

Out o

isInput :: IOAct i o -> Bool

Is the given action an input?

isOutput :: IOAct i o -> Bool

Is the given action an output?

fromInput :: IOAct i o -> i

Partially defined function that unpacks an input.

maybeFromInput :: IOAct i o -> Maybe i

Unpacks an input.

fromOutput :: IOAct i o -> o

Partially defined function that unpacks an outputs.

maybeFromOutput :: IOAct i o -> Maybe o

Unpacks an output.

Observable Quiescence

Observable actions that may be either inputs provided to a system, or outputs from that system, where the output may also be an artificial output that represents an observed timeout, quiescence. For the theoretical background on observing quiescence, see

Jan Tretmans, Model based testing with labelled transition systems (Formal Methods and Testing), 2008

data Suspended o

Add observation of quiescence to a type of observable actions.

Constructors

Quiescence

OutSusp o

δ :: IOAct i (Suspended o)

type IOSuspAct i o = IOAct i (Suspended o)

Add observation of quiescence to the observed inputs and outputs.

asSuspended :: IOAct i o -> IOSuspAct i o

Convert an input or output to a type containing quiescence.

fromSuspended :: IOSuspAct i o -> IOAct i o

Partially defined function that unpacks an input or output from a type with quiescence.

Input Failures

'Input failures' model the possibility to accepted or refuse an input by the system under test. Failure of an input represents the system being unable to process the given input. For a theoretical background on input failures, see

Ramon Janssen, Refinement and partiality for model-based testing (Doctoral dissertation), 2022, Chapter 3 and 4

type IFAct i o = IOAct (InputAttempt i) o

Observable actions that may be either inputs provided to a system, or outputs from that system, where the inputs may be refused.

newtype InputAttempt i

Constructors

InputAttempt (i, Bool)

asInputAttempt :: IOAct i o -> IFAct i o

Convert an input or output to a type containing input failures.

fromInputAttempt :: IFAct i o -> IOAct i o

Partially defined function that unpacks an input or output from a type with input failures.

Combined Input Failures and Quiescences

type SuspendedIF i o = IOAct (InputAttempt i) (Suspended o)

Input failure with observed quiescence. See IOSuspAct and IFAct for details.

asSuspendedInputAttempt :: IOAct i o -> SuspendedIF i o

Convert an input or output to a type containing input failures and quiescence.

fromSuspendedInputAttempt :: SuspendedIF i o -> IOAct i o

Partially defined function that unpacks an input or output from a type with input failures and quiescence.

STS

data SymInteract g

Constructors

SymInteract g [Variable]

type IOSymInteract i o = SymInteract (IOAct i o)

type SymGuard = Expr Bool

data GateValue g

Constructors

GateValue g [Constant]

type IOGateValue i o = GateValue (IOAct i o)

gateValueAsIOAct :: IOGateValue i o -> IOAct (GateValue i) (GateValue o)

ioActAsGateValue :: IOAct (GateValue i) (GateValue o) -> IOGateValue i o

maybeFromInputInteraction :: IOSymInteract i o -> Maybe (SymInteract i)

maybeFromOutputInteraction :: IOSymInteract i o -> Maybe (SymInteract o)

isInputGate :: IOGateValue i o -> Bool

isOutputGate :: IOGateValue i o -> Bool

isInputInteract :: IOSymInteract i o -> Bool

isOutputInteract :: IOSymInteract i o -> Bool

interactionGate :: SymInteract g -> g

valueGate :: GateValue g -> g

type IOSuspGateValue i o = IOGateValue i (Suspended o)

type IFGateValue i o = IOGateValue (InputAttempt i) o

type SuspendedIFGateValue i o = IOGateValue (InputAttempt i) (Suspended o)

toIOGateValue :: SuspendedIF (GateValue i) (GateValue o) -> SuspendedIFGateValue i o