Lattest.Model.BoundedMonad

Description

A bounded monad is a type constructor which represents the observable perspective on the state of an automaton, also called a state configuration. A system will internally have a state, but this state can not always be observed or inferred for an external viewer. To model the difference between internal and observable state, we define automata as having a type of internal state, say q, and an observable state configuration over q.

In this module, we define three such state configurations:

deterministic state configurations, where every behaviour leads to a single state,

non-deterministic state configurations, where given observable behaviour may lead to a set of states, and

distributive lattices, or positive boolean formulas, where the observable behaviour is expressed as a logical expression over states.

Here, 'observed behaviour' is formally a trace or sequence of observable actions. Thus, after a trace, the system is in a state configuration over states. The state configuration is computed by taking transitions, following the monadic interpret of the state configuration. If the automaton does not exhibit a given trace of actions, then the state configuration after that trace is either forbidden or underspecified. Here, forbidden expresses that the automaton does not allow the given trace, whereas underspecified expresses that the automaton does not specify the trace, hence the trace is allowed.

The name 'bounded monad' is based on bounded lattices, which have a top and bottom element. For more details on the use of top and bottom in state configurations, see

Ramon Janssen, Refinement and partiality for model-based testing (Doctoral dissertation), 2022, Chapter 4

State configurations

Deterministic

data Det q

Deterministic state configuration. This means that an automaton is either in a single state, or in an explicit forbidden configuration, or in an explicit underspecified configuration.

Constructors

Det q

ForbiddenDet

UnderspecDet

Non-deterministic

data NonDet q

Non-deterministic state configuration. This means that an automaton non-deterministically in a number of states, where zero states indicates the forbidden configuration, or in an explicit underspecified configuration.

Constructors

NonDet [q]

UnderspecNonDet

Distributive lattice

data FreeLattice a

Free distributive lattice, or a positive boolean formula, i.e., a boolean formula with conjunctions and disjunctions over atomic propositions. The two elements top and bot can be interpreted as true and false.

atom :: a -> FreeLattice a

A single state embedded in a free distributive lattice.

top :: FreeLattice a

The free distributive lattice element ⊤, or true.

bot :: FreeLattice a

The free distributive lattice element ⊥, or false.

(/\) :: FreeLattice a -> FreeLattice a -> FreeLattice a

Conjunction on free distributive lattices.

Specifiednesss

data Specifiedness

Specifiednesss describe wether behaviour (a sequence of actions) is allowed a stateful specification model. Forbidden describes that behaviour is not allowed, not is any subsequent behaviour. Underspecified describes that behaviour is allowed and any subsequent behaviour is also allowed. Indefinite means that the behaviour is allowed, and subsequent behaviour may be forbidden, underspecified, or may again be Indefinite.

Constructors

Underspecified

Forbidden

Indefinite

class BoundedConfiguration (m :: k -> Type)

Specifiedness configurations are state configurations which have a representation for forbidden and underspecified configurations.

Minimal complete definition

forbidden, underspecified, isForbidden, isUnderspecified

specifiedness :: forall {k} {m} {t :: k}. BoundedConfiguration m => m t -> Specifiedness

Extract the current Specifiedness from a bounded monad.

forbidden

Arguments

:: forall (t :: k). BoundedConfiguration m

=> m t
The forbidden state configuration.

underspecified

Arguments

:: forall (t :: k). BoundedConfiguration m

=> m t
The underspecified state configuration.

isForbidden

Arguments

:: forall (t :: k). BoundedConfiguration m

=> m t

-> Bool
Is this state configuration forbidden?

isUnderspecified

Arguments

:: forall (t :: k). BoundedConfiguration m

=> m t

-> Bool
Is this state configuration underspecified?

Auxiliary Specifiedness functions

isAllowed :: forall {k} m (t :: k). BoundedConfiguration m => m t -> Bool

Is the configuration a representation of "allowed", i.e., Indefinite or Underspecified?

isSpecified :: forall {k} m (t :: k). BoundedConfiguration m => m t -> Bool

Is the configuration a representation of "specified", i.e., Indefinite or Forbidden?

isIndefinite :: forall {k} m (t :: k). BoundedConfiguration m => m t -> Bool

Is the configuration a representation of the Indefinite specifiedness?

isConclusive :: forall {k} m (t :: k). BoundedConfiguration m => m t -> Bool

Is the configuration a representation of "definitive", i.e., Forbidden or Underspecified?

Utility types

type BoundedMonad (m :: Type -> Type) = (BoundedConfiguration m, Monad m)

Abbreviation for types which are both bounded configurations and Monads.

type BoundedApplicative (m :: Type -> Type) = (BoundedConfiguration m, Applicative m)

Abbreviation for types which are both bounded configurations and Applicatives.

type BoundedFunctor (m :: Type -> Type) = (BoundedConfiguration m, Functor m)

Abbreviation for types which are both bounded configurations and Functors.

General non-determinism

class JoinSemiLattice a

Because the lattices-library doesn't support this

Minimal complete definition

(\/)

(\/) :: JoinSemiLattice a => a -> a -> a