Lattest.Model.StandardAutomata

Description

This module contains some simple automata types, and auxiliary functions for constructing them in a convenient manner. Note that the types are no more than type aliases, so if the automata type with your favorite combination of type parameters is not listed, don't hesitate to construct an automaton type (AutSyn or AutSem) yourself instead of using the StandardAutomata.

Syntactical automata

Standard automata types

type IA m loc i o = AutSyn m loc (IOAct i o) ()

Interface Automaton (for now, without internal transitions)

type DIA loc i o = IA DetState loc i o

deterministic IA

type NDIA loc i o = IA NonDetState loc i o

non-deterministic IA

type AIA loc i o = IA FDL loc i o

alternating IA

Auxiliary Functions for Constructing Automata

Interface Automata

Functions for constructing Interface Automata, with arbitrary state configurations.

ia :: Functor m => m loc -> (loc -> Map (IOAct i o) (m loc)) -> IA m loc i o

Construct an IA from transitions expressed as maps from actions to state configurations.

iaWithMenu :: (PermissionFunctor m, Foldable fld, Ord i, Ord o) => m loc -> (loc -> fld (IOAct i o)) -> (loc -> Map (IOAct i o) (m loc)) -> IA m loc i o

Construct an IA from transitions expressed as maps from actions to state configurations. Also accepts a menu of actions that defines, per state, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit map of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

iaWithAlphabet :: (PermissionFunctor m, Foldable fld, Ord i, Ord o) => m loc -> fld (IOAct i o) -> (loc -> Map (IOAct i o) (m loc)) -> IA m loc i o

Construct an IA from transitions expressed as maps from actions to state configurations. Also accepts an alphabet of actions that defines, globally for the entire automaton, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit map of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

Deterministic Interface Automata

dia :: loc -> (loc -> Map (IOAct i o) (Maybe loc)) -> DIA loc i o

Construct a deterministic IA from transitions expressed as maps from actions to Just states, or Nothing for actions that are not enabled.

diaWithMenu :: (Ord i, Ord o, Foldable fld) => loc -> (loc -> fld (IOAct i o)) -> (loc -> Map (IOAct i o) (Maybe loc)) -> DIA loc i o

Construct a deterministic IA from transitions expressed as maps from actions to Just states. Also accepts a menu of actions that defines, per state, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit map of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

diaWithAlphabet :: (Ord i, Ord o, Foldable fld) => loc -> fld (IOAct i o) -> (loc -> Map (IOAct i o) (Maybe loc)) -> DIA loc i o

Construct a deterministic IA from transitions expressed as maps from actions to Just states. Also accepts an alphabet of actions that defines, globally for the entire automaton, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit map of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

dia' :: (Ord loc, Ord i, Ord o) => loc -> [(loc, IOAct i o, loc)] -> DIA loc i o

Construct a deterministic IA from a list of transitions

diaWithMenu' :: (Ord loc, Ord i, Ord o, Foldable fld) => loc -> (loc -> fld (IOAct i o)) -> [(loc, IOAct i o, loc)] -> DIA loc i o

Construct a deterministic IA from a list of transitions. Also accepts a menu of actions that defines, per state, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit list of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

diaWithAlphabet' :: (Ord loc, Ord i, Ord o, Foldable fld) => loc -> fld (IOAct i o) -> [(loc, IOAct i o, loc)] -> DIA loc i o

Construct a deterministic IA from a list of transitions. Also accepts an alphabet of actions that defines, globally for the entire automaton, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit list of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

Non-Deterministic Interface Automata

ndia :: [loc] -> (loc -> Map (IOAct i o) [loc]) -> NDIA loc i o

Construct an non-deterministic IA from transitions expressed as maps from actions to lists of states, or the empty list for actions that are not enabled.

ndiaWithMenu :: (Ord i, Ord o, Foldable fld) => [loc] -> (loc -> fld (IOAct i o)) -> (loc -> Map (IOAct i o) [loc]) -> NDIA loc i o

Construct an non-deterministic IA from transitions expressed as maps from actions to lists of states. Also accepts a menu of actions that defines, per state, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit map of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

ndiaWithAlphabet :: (Ord i, Ord o, Foldable fld) => [loc] -> fld (IOAct i o) -> (loc -> Map (IOAct i o) [loc]) -> NDIA loc i o

Construct an non-deterministic IA from transitions expressed as maps from actions to lists of states. Also accepts an alphabet of actions that defines, globally for the entire automaton, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit map of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

Alternating Interface Automata

aia :: FDL loc -> (loc -> Map (IOAct i o) (FDL loc)) -> AIA loc i o

Construct an alternating IA from transitions expressed as maps from actions to logical expressions over states, or top or bottom for actions that are not enabled.

aiaWithMenu :: (Foldable fld, Ord i, Ord o) => FDL loc -> (loc -> fld (IOAct i o)) -> (loc -> Map (IOAct i o) (FDL loc)) -> AIA loc i o

Construct an alternating IA from transitions expressed as maps from actions to logical expressions over states. Also accepts a menu of actions that defines, per state, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit map of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

aiaWithMenu' :: (Foldable fld, Ord i, Ord o, Ord loc) => FDL loc -> (loc -> fld (IOAct i o)) -> [(loc, IOAct i o, FDL loc)] -> IA FDL loc i o

Construct an alternating IA from a list of transitions to logical expressions over states. Also accepts a menu of actions that defines, per state, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit map of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

aiaWithAlphabet :: (Foldable fld, Ord i, Ord o) => FDL loc -> fld (IOAct i o) -> (loc -> Map (IOAct i o) (FDL loc)) -> AIA loc i o

Construct an alternating IA from transitions expressed as maps from actions to logical expressions over states. Also accepts an alphabet of actions that defines, globally for the entire automaton, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit map of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

aiaWithAlphabet' :: (Foldable fld, Ord i, Ord o, Ord loc) => FDL loc -> fld (IOAct i o) -> [(loc, IOAct i o, FDL loc)] -> AIA loc i o

Construct an alternating IA from a list of transitions to logical expressions over states. Also accepts an alphabet of actions that defines, globally for the entire automaton, which actions are (at least) implicitly defined. Actions that are in the menu but not in the explicit map of transitions are mapped to underspecified (for inputs) or forbidden (for outputs).

Alternating state configurations

data FDL a

Free distributive lattice, or a positive boolean formula, i.e., a boolean formula with conjunctions and disjunctions over atomic propositions. The two elements top and bot can be interpreted as true and false.

atom :: a -> FDL a

A single state embedded in a free distributive lattice.

top :: FDL a

The free distributive lattice element ⊤, or true.

bot :: FDL a

The free distributive lattice element ⊥, or false.

(\/) :: FDL a -> FDL a -> FDL a

Disjunction on free distributive lattices.

(/\) :: FDL a -> FDL a -> FDL a

Conjunction on free distributive lattices.

Automaton Semantics

type ConcreteAutSem m q act = AutSem m q q act () act

Semantics of automata in which syntactical states and actions are directly interpreted as literal, semantical states and actions.

semanticsConcrete :: (StateConfiguration m, Ord t) => AutSyn m loc t () -> ConcreteAutSem m loc t

Interpret syntactical states and actions directly as literal, semantical states and actions.

type ConcreteTimeoutAutSem m q i o = AutSem m q q (IOAct i o) () (TimeoutIO i o)

Semantics of automata in which syntactical states and actions are directly interpreted as literal, semantical states and actions, but with timeouts as possible output observations.

semanticsQuiescentConcrete :: (StateConfiguration m, Ord i, Ord o) => AutSyn m loc (IOAct i o) () -> ConcreteTimeoutAutSem m loc i o

Interpret syntactical states and actions are directly as literal, semantical states and actions, but with timeouts as possible output observations.

type ConcreteTimeoutInputAttemptAutSem m q i o = AutSem m q q (IOAct i o) () (TimeoutIF i o)

Semantics of automata in which syntactical states and actions are directly interpreted as literal, semantical states and actions, but with timeouts and input failures as possible observations.

semanticsQuiescentInputAttemptConcrete :: (StateConfiguration m, Ord i, Ord o) => AutSyn m loc (IOAct i o) () -> ConcreteTimeoutInputAttemptAutSem m loc i o

Interpret syntactical states and actions are directly as literal, semantical states and actions, but with timeouts and input failures as possible observations.