Lattest.Model.StateConfiguration

Description

A state configuration is a type constructor which represents the observable perspective on the state of an automaton. A system will internally have a state, but this state can not always be observed or inferred for an external viewer. To model the difference between internal and observable state, we define automata as having a type of internal state, say q, and an observable state configuration over q.

In this module, we define three such state configurations:

deterministic state configurations, where every behaviour leads to a single state,

non-deterministic state configurations, where given observable behaviour may lead to a set of states, and

distributive lattices, or positive boolean formulas, where the observable behaviour is expressed as a logical expression over states.

Here, 'observed behaviour' is formally a trace or sequence of observable actions. Thus, after a trace, the system is in a state configuration over states. The state configuration is computed by taking transitions, following the monadic semantics of the state configuration. If the automaton does not exhibit a given trace of actions, then the state configuration after that trace is either forbidden or underspecified. Here, forbidden expresses that the automaton does not allow the given trace, whereas underspecified expresses that the automaton does not specify the trace, hence the trace is allowed.

For more details see

Ramon Janssen, Refinement and partiality for model-based testing (Doctoral dissertation), 2022, Chapter 4

State configurations

Deterministic

data DetState q

Deterministic state configuration. This means that an automaton is either in a single state, or in an explicit forbidden configuration, or in an explicit underspecified configuration.

Constructors

Det q

ForbiddenDet

UnderspecDet

Non-deterministic

data NonDetState q

Non-deterministic state configuration. This means that an automaton non-deterministically in a number of states, where zero states indicates the forbidden configuration, or in an explicit underspecified configuration.

Constructors

NonDet [q]

UnderspecNonDet

Distributive lattice

data FDL a

Free distributive lattice, or a positive boolean formula, i.e., a boolean formula with conjunctions and disjunctions over atomic propositions. The two elements top and bot can be interpreted as true and false.

atom :: a -> FDL a

A single state embedded in a free distributive lattice.

top :: FDL a

The free distributive lattice element ⊤, or true.

bot :: FDL a

The free distributive lattice element ⊥, or false.

(\/) :: FDL a -> FDL a -> FDL a

Disjunction on free distributive lattices.

(/\) :: FDL a -> FDL a -> FDL a

Conjunction on free distributive lattices.

Permissions

data Permission

Permissions describe wether behaviour (a sequence of actions) is allowed a stateful specification model. Forbidden describes that behaviour is not allowed, not is any subsequent behaviour. Underspecified describes that behaviour is allowed and any subsequent behaviour is also allowed. Indefinite means that the behaviour is allowed, and subsequent behaviour may be forbidden, underspecified, or may again be Indefinite.

Constructors

Underspecified

Forbidden

Indefinite

class PermissionConfiguration m

Permission configurations are state configurations which have a representation for forbidden and underspecified configurations.

Minimal complete definition

forbidden, underspecified, isForbidden, isUnderspecified

permission :: forall {k} {m} {t :: k}. PermissionConfiguration m => m t -> Permission

Extract the current Permission from a permission configuration.

forbidden

Arguments

:: PermissionConfiguration m

=> m t
The forbidden state configuration.

underspecified

Arguments

:: PermissionConfiguration m

=> m t
The underspecified state configuration.

isForbidden

Arguments

:: PermissionConfiguration m

=> m t

-> Bool
Is this state configuration forbidden?

isUnderspecified

Arguments

:: PermissionConfiguration m

=> m t

-> Bool
Is this state configuration underspecified?

Auxiliary Permission functions

isAllowed :: PermissionConfiguration m => m t -> Bool

Is the configuration a representation of an "allowed" permission, i.e., Indefinite or Underspecified?

isSpecified :: PermissionConfiguration m => m t -> Bool

Is the configuration a representation of a "specified" permission, i.e., Indefinite or Forbidden?

isIndefinite :: PermissionConfiguration m => m t -> Bool

Is the configuration a representation of the Indefinite permission?

isConclusive :: PermissionConfiguration m => m t -> Bool

Is the configuration a representation of a "definitive" permission, i.e., Forbidden or Underspecified?

Utility types

type StateConfiguration m = (PermissionConfiguration m, Monad m)

Abbreviation for types which are both permission configurations and Monads.

type PermissionApplicative m = (PermissionConfiguration m, Applicative m)

Abbreviation for types which are both permission configurations and Applicatives.

type PermissionFunctor m = (PermissionConfiguration m, Functor m)

Abbreviation for types which are both permission configurations and Functors.