Lattest.Model.Symbolic

Description

This module contains the definitions for working with symbolic values: variables, expressions, concrete values, guards and assignments.

Documentation

type SymGuard = SymBool

A symbolic guard, i.e., an predicate.

type SymAssign = Map Variable SymExpr

Assignments of symbolic expressions to variables.

data Value

Typed, concrete values.

Constructors

IntVal Integer

BoolVal Bool

data Variable

A typed variable.

Constructors

Variable Symbol Type

addTypedVar :: Variable -> Value -> Model -> Model

In a symbolic model, substitute a variable by an expression.

data Type

Types of symbolic expressions.

Note: the supported types will be extended in the future.

Constructors

IntType

BoolType

data SymExpr

A typed expression.

Constructors

BoolExpr SymBool

IntExpr SymInteger

equalTyped :: Variable -> Value -> Bool

Check whether the types of a variable and an expression match.

type Valuation = Map Variable Value

Assignments of concrete expressions to variables.