Lattest.Model.Symbolic.Expr

Documentation

data Expr t

Expr: value expression 1. User can't directly construct Expr (such that invariants will always hold) 2. User can still pattern match on Expr using ExprView 3. Overhead at run-time is zero. See https://wiki.haskell.org/Performance/Data_types#Newtypes

view :: Expr t -> ExprView t

data ExprView t where

Constructors

Var

Fields

:: forall t. { variable :: Variable

} -> ExprView t

Const

Fields

:: forall t. ExprType t

=> { constant :: t

} -> ExprView t

Ite

Fields

:: forall t. { conditional :: ExprView Bool

, trueBranch :: ExprView t

, falseBranch :: ExprView t

} -> ExprView t

EqualInt

Fields

:: { leftInt :: ExprView Integer

, rightInt :: ExprView Integer

} -> ExprView Bool

EqualString

Fields

:: { leftString :: ExprView String

, rightString :: ExprView String

} -> ExprView Bool

EqualBool

Fields

:: { leftBool :: ExprView Bool

, rightBool :: ExprView Bool

} -> ExprView Bool

Divide

Fields

:: { dividend2 :: ExprView Integer

, divisor2 :: ExprView Integer

} -> ExprView Integer

Modulo

Fields

:: { dividend2 :: ExprView Integer

, divisor2 :: ExprView Integer

} -> ExprView Integer

Sum :: FreeSum (ExprView Integer) -> ExprView Integer

Product :: FreeProduct (ExprView Integer) -> ExprView Integer

Length :: ExprView String -> ExprView Integer

GezInt :: ExprView Integer -> ExprView Bool

Not :: ExprView Bool -> ExprView Bool

And :: Set (ExprView Bool) -> ExprView Bool

At

Fields

:: { string2 :: ExprView String

, position2 :: ExprView Integer

} -> ExprView [Char]

Concat :: [ExprView String] -> ExprView [Char]

data Type

Constructors

IntType

BoolType

StringType

varName :: Variable -> String

varType :: Variable -> Type

data Constant

Constructors

Cbool
Constructor of Boolean constant.

Fields

toBool :: Bool

Cint
Constructor of Integer constant.

Fields

toInteger :: Integer

Cstring
Constructor of String constant.

Fields

toString :: String

Ccstr
Constructor of constructor constant (value of ADT).

Fields

cstrName :: String

args :: [Constant]

constType :: Constant -> Type

class ConstType t

convert a Constant to an typed value

Minimal complete definition

fromConst, toConst

fromConst :: ConstType t => Constant -> Either String t

toConst :: ConstType t => t -> Constant

data Variable

Constructors

Variable

Fields

varName :: String

varType :: Type

data VarModel

class Assignable t

Minimal complete definition

assign, assignValue, assignedExpr, assignedExprWithDefault

assign :: Assignable t => Variable -> Expr t -> VarModel -> VarModel

assignValues :: [Valuation -> Valuation] -> Valuation

eval :: Expr v -> Either String v

Evaluate the provided value expression. Either the Right Constant Value is returned or a (Left) error message.

subst

Arguments

:: Assignable t

=> VarModel
Map from variables to value expressions.

-> Expr t
Value expression where the substitution will take place.

-> Expr t

Substitute variables by value expressions in a value expression.

Preconditions are not checked.

substConst :: Assignable t => Valuation -> Expr t -> Expr t

data Valuation

assignedExpr :: Assignable t => Variable -> VarModel -> Maybe (Expr t)

(=:) :: Assignable t => Variable -> Expr t -> VarModel -> VarModel infixr 0

freeVars :: Expr t -> Set Variable

module Lattest.Model.Symbolic.Internal.ExprImpls

module Lattest.Model.Symbolic.Internal.ExprImplsExtension