Lattest.Model.Symbolic.Internal.ExprImpls

Constructors to create Value Expressions

Constant value

sConst :: ExprType t => t -> Expr t

Create a function call. Preconditions are not checked.

Is Expr a Constant/Value Expression? isConst :: Expr -> Bool isConst (view -> Vconst{}) = True isConst _ = False

sTrue :: Expr Bool

sFalse :: Expr Bool

VarRef

sVar :: VarExpr t => Variable -> Expr t

General Operators to create Value Expressions

Equal

(.==) :: EqExpr t => Expr t -> Expr t -> Expr Bool infix 4

If Then Else

sIfThenElse :: Expr Bool -> Expr t -> Expr t -> Expr t

Apply operator ITE (IF THEN ELSE) on the provided value expressions. Preconditions are not checked.

Function Call

Boolean Operators to create Value Expressions

Not

sNot :: Expr Bool -> Expr Bool

Apply operator Not on the provided value expression. Preconditions are not checked.

And

sAnd :: Set (Expr Bool) -> Expr Bool

Apply operator And on the provided set of value expressions. Preconditions are not checked.

Integer Operators to create Value Expressions

Sum

sSum :: FreeSum (Expr Integer) -> Expr Integer

Product

sProduct :: FreeProduct (Expr Integer) -> Expr Integer

Divide

(./) :: Expr Integer -> Expr Integer -> Expr Integer infixl 7

Apply operator Divide on the provided value expressions. Preconditions are not checked.

Modulo

(.%) :: Expr Integer -> Expr Integer -> Expr Integer infixl 7

Apply operator Modulo on the provided value expressions. Preconditions are not checked.

Comparisons GEZ

sIsNonNegative :: Expr Integer -> Expr Bool

Apply operator GEZ (Greater Equal Zero) on the provided value expression. Preconditions are not checked.

String Operators to create Value Expressions

Length operator

sLength :: Expr String -> Expr Integer

Apply operator Length on the provided value expression. Preconditions are not checked.

At operator

(.@) :: Expr String -> Expr Integer -> Expr String infixl 5

Apply operator At on the provided value expressions. Preconditions are not checked.

Concat operator

sConcat :: [Expr String] -> Expr String

Apply operator Concat on the provided sequence of value expressions. Preconditions are not checked.

Regular Expression Operators to create Value Expressions

String in Regular Expression operator

Algebraic Data Type Operators to create Value Expressions

Algebraic Data Type constructor operator

Algebraic Data Type IsConstructor function

Algebraic Data Type Accessor

Substitution of var by value

data VarModel

class Assignable t

Minimal complete definition

assign, assignValue, assignedExpr, assignedExprWithDefault

assign :: Assignable t => Variable -> Expr t -> VarModel -> VarModel

data Valuation

toConstantsMap :: Valuation -> Map Variable Constant

fromConstantsMap :: Map Variable Constant -> Valuation

emptyValuation :: Valuation

assignValues :: [Valuation -> Valuation] -> Valuation

assignValue :: Assignable t => Variable -> t -> Valuation -> Valuation

insertIntoValuation :: Variable -> Constant -> Valuation -> Valuation

substConst :: Assignable t => Valuation -> Expr t -> Expr t

subst

Arguments

:: Assignable t

=> VarModel
Map from variables to value expressions.

-> Expr t
Value expression where the substitution will take place.

-> Expr t

Substitute variables by value expressions in a value expression.

Preconditions are not checked.

assignedExpr :: Assignable t => Variable -> VarModel -> Maybe (Expr t)

assignment :: [VarModel -> VarModel] -> VarModel

noAssignment :: VarModel

(=:) :: Assignable t => Variable -> Expr t -> VarModel -> VarModel infixr 0