Lattest.Model.Symbolic.Internal.FreeMonoidX

Description

a0 <> a1 <> ... <> an-1

`ai in A`, <> is a commutative (binary) operation, `(A, <>)` is a Monoid, there is a multiplication operator <.> that is distributive w.r.t <>, and there is additive inverse for the elements of A:

a <> b         = b <> a                 -- for all a, b in A (commutativity)
a <.> (b <> c) = (a <.> b) <> (a <.> c) -- for all a, b, c in A (left distributivity).
(b <> c) <.> a = (b <.> a) <> (c <.> a) -- for all a, b, c in A (right distributivity).
a <> -a        = 0                      -- for all a, for some (-a), where `0 = mempty`.

Furthermore, to be able to assign semantics to this symbolic representation it is required that the terms are integer-multipliable, meaning that

`a <.> n`

is equivalent to adding an element a n times if n is non-negative, or removing it n times otherwise. See IntMultipliable class, and multiplyLaw in FreeMonoidXSpec.

Note that we're using <.> as having both types `A -> A` and `Integral n => n -> a`, which is fine at the moment since the FreeMonoidX's we are dealing with are numeric types.

`FreeMonoidX a` is an instance of the IsList class, which means that in combination with the OverloadedLists extension is possible to write free-monoids as lists. For instance the free-monoid:

a0 <> a1 <> ... <> an-1

Can be written as

a0, a1, ..., an-1

Free Monoid with multiplication Type

newtype FreeMonoidX a

Symbolic representation of a polynomial, where each term is a member of type a.

The integer value of the map represents the number of occurrences of a term in the polynomial. Given this representation it is crucial that the operation is commutative, since the information of about the order of the term is lost in this representation.

Constructors

FMX

Fields

asMap :: Map a Integer

Query

nrofDistinctTerms :: FreeMonoidX a -> Int

Number of distinct terms in the free-monoid.

distinctTerms :: FreeMonoidX a -> [a]

O(n). The distinct terms of a free-monoid., each term occurs only once in the list.

distinctTermsT :: TermWrapper t => FreeMonoidX (t a) -> [a]

allFreeMonoidX :: (TermWrapper f, Functor f) => (a -> Bool) -> FreeMonoidX (f a) -> Bool

Map

mapTerms :: Ord b => (a -> b) -> FreeMonoidX a -> FreeMonoidX b

Map the terms of the free-monoid.

mapFreeMonoidX :: (TermWrapper f, Functor f, Ord (f b)) => (a -> b) -> FreeMonoidX (f a) -> FreeMonoidX (f b)

Filter

partition :: (a -> Bool) -> FreeMonoidX a -> (FreeMonoidX a, FreeMonoidX a)

O(n). Partition the free-monoid into two free-monoids, one with all elements that satisfy the predicate and one with all elements that don't satisfy the predicate.

partitionT :: TermWrapper t => (a -> Bool) -> FreeMonoidX (t a) -> (FreeMonoidX (t a), FreeMonoidX (t a))

Folds

foldrTerms :: (TermWrapper f, Foldable f) => (a -> b -> b) -> b -> FreeMonoidX (f a) -> b

foldOccur :: (a -> Integer -> b -> b) -> b -> FreeMonoidX a -> b

O(n). Fold over the terms of the free-monoid with their multipliers.

foldFMX :: (IntMultipliable a, Monoid a, Semigroup a) => FreeMonoidX a -> a

Fold the free-monoid.

fold :: (IntMultipliable (t a), Monoid (t a), TermWrapper t) => FreeMonoidX (t a) -> a

Fold the free-monoid of term wrappers.

Manipulation of the free monoid

append :: Ord a => a -> FreeMonoidX a -> FreeMonoidX a

Append a term to the free-monoid.

append 2 [1, 2, 3]

should be equivalent to the free-monoid.

[1, 2, 3, 2]

remove :: Ord a => a -> FreeMonoidX a -> FreeMonoidX a

Remove a term from the free-monoid.

remove 2 (1 <> 2 <> 3)

should be equivalent to the free-monoid

(1 <> 3)

appendMany :: Ord a => Integer -> a -> FreeMonoidX a -> FreeMonoidX a

Add the term x n times. If n is negative the term will be removed n times.

appendMany 2 10 (10 <> 12 <> 12)

should be equivalent to

(10 <> 10 <> 10 <> 12 <> 12)

appendMany (-2) 10 (10 <> 12 <> 12)

should be equivalent to

(-10 <> 12 <> 12)

flatten :: Ord a => FreeMonoidX (FreeMonoidX a) -> FreeMonoidX a

Flatten a free-monoid.

For instance, the monoid

(a <> b) <> (a <> b) <> a

will be rewritten as:

a <> a <> a <> b <> b

Assuming `a < b`.

Multiplication operator

(<.>)

Arguments

:: (IntMultipliable a, Integral n)

=> n
Multiplication factor.

-> a
Element to multiply.

-> a

`n . x` multiplies x n times.

Monoid terms restriction

class IntMultipliable a

Types that can be multiplied by an integral. This restriction is required to be able to assign a correct semantic to the symbolic representation of the free-monoid. If the elements of the free-monoid are not IntMultipliable, then foldFMX is ill-defined.

See the test code FreeMonoidXspec for examples of instances of this class.

Minimal complete definition

(<.>)

class TermWrapper (f :: Type -> Type) where

A term of the monoid which wraps a value. This could be for instance a sum-term or a product term.

Methods

wrap :: a -> f a

unwrap :: f a -> a

Occurrence lists

toOccurList :: FreeMonoidX a -> [(a, Integer)]

O(n). Convert the free-monoid to a list of (term, occurrence) tuples.

For instance given the term:

a <> b <> a <> b <> a

toOccurList applied to it will result in the following list:

[(a, 3), (b, 2)]

toDistinctAscOccurList :: FreeMonoidX a -> [(a, Integer)]

O(n). Convert the free-monoid to a distinct ascending list of term/multiplier pairs.

fromOccurList :: Ord a => [(a, Integer)] -> FreeMonoidX a

O(n*log n). Create a free-monoid from a list of term/multiplier pairs.

fromDistinctAscOccurList :: [(a, Integer)] -> FreeMonoidX a

O(n). Build a free-monoid from an ascending list of term/multiplier pairs where each term appears only once. /The precondition (input list is strictly ascending) is not checked./

Occurrence lists of TermWrapper's

fromListT :: (Ord (t a), TermWrapper t) => [a] -> FreeMonoidX (t a)

O(t*log t). Create a product from a list of terms using the given term wrapper.

toOccurListT :: TermWrapper t => FreeMonoidX (t a) -> [(a, Integer)]

toDistinctAscOccurListT :: TermWrapper t => FreeMonoidX (t a) -> [(a, Integer)]

fromOccurListT :: (Ord (t a), TermWrapper t) => [(a, Integer)] -> FreeMonoidX (t a)

fromDistinctAscPowerListT :: TermWrapper t => [(a, Integer)] -> FreeMonoidX (t a)

Lists of direct items

toList :: IsList l => l -> [Item l]

fromList :: IsList l => [Item l] -> l