Lattest.Model.Symbolic.SolveSTS

Description

Find concrete values to take transitions in STSes, using an SMT solver.

Documentation

solveRandomInteraction :: forall (m :: Type -> Type) g r loc g'' g'. (OrdMonad m, BooleanConfiguration m, Ord g, Ord (m (Expr Bool)), RandomGen r) => AutIntrpr m loc (IntrpState loc) (SymInteract g) STStdest (GateValue g'') -> (SymInteract g -> Maybe (SymInteract g')) -> r -> SMT (Maybe (GateValue g'), r)

For the given STS and a subset function, using SMT solving, find a interaction of the STS in that subset for which the guard is true from the current STS state. The interaction is picked uniformly randomly among interactions with satisfied gates, if any. This uses the supplied random generator and returns the new random generator state. The returned gate values for that interaction are not randomized in any way, picking values is left to the SMT solver.