Lattest.Model.Symbolic.SolveSymPrim

Documentation

combineGuards :: (BooleanConfiguration m, OrdFunctor m) => m SymGuard -> SymGuard

Combine the given guards into one.

substituteInGuard :: Valuation -> SymGuard -> SymGuard

In the given guard, substitute the given valuation.

evaluateGuard :: SymGuard -> Bool

Evaluate the given guard

solveAnySequential :: [(SymInteract g, SymGuard)] -> SMT (Maybe (GateValue g))

For the given list of interactions and guards, using SMT solving, pick the first interaction in that list for which the guard is satisfiable, if any. The returned gate values for that interaction are not randomized in any way, picking values is left to the SMT solver.

solveGuard :: [Variable] -> SymGuard -> SMT (Maybe Valuation)