Lattest.SMT.Config

Description

This module defines the options that can be passed to TorXakis.

Documentation

data Config

TorXakis configuration options.

Constructors

Config

Fields

smtLog :: Bool

Log all SMT commands?

availableSolvers :: Map SolverId SolverConfig

Available solvers that can be chosen from.

selectedSolver :: SolverId

configuredParameters :: [(ParamName, ParamValue)]

getProc :: Config -> Maybe CreateProcess

Return a CreateProcess according to the selected solver in the configuration.

defaultConfig :: Config

TorXakis default configuration

newtype SolverId

Constructors

SolverId

Fields

solverId :: String

data SolverConfig

Constructors

SolverConfig

Fields

execName :: FilePath

Executable name.

smtArgs :: [String]

Arguments for to be passed to the smtSolver.

newtype ParamName

Constructors

ParamName String

newtype ParamValue

Constructors

ParamValue String

changeSolver :: Config -> String -> Config

changeLog :: Config -> Bool -> Config

addSolvers :: Config -> Map SolverId SolverConfig -> Config

setParams :: Config -> [(ParamName, ParamValue)] -> Config

update values in a Map of Params with a list of configured parameters

updateParamVals

Arguments

:: Params
Map of parameters ( Map.Map String (String,String->Bool) )

-> [(ParamName, ParamValue)]
List of configured parameters

-> Params

Updates a Map of parameters' values with given list of configured parameter

updateCfg :: Maybe a -> (Config -> a -> Config) -> Config -> Config