Lattest.Util.IOUtils

Description

An unsorted collection of utility functions for working with the IO monad.

Control flow

ifM_ :: Applicative m => Bool -> m a -> m ()

If-statement for monads, with a pure condition: if the condition holds, execute the given monad and discard the result, otherwise return ().

ifMM_ :: Monad m => m Bool -> m a -> m ()

If-statement for monads, with a monadic condition: if the condition holds, execute the given monad and discard the result, otherwise return ().

ifM :: Applicative m => Bool -> m a -> m (Maybe a)

If-statement for monads, with a pure condition: if the condition holds, execute the given monad and return the result, otherwise return Nothing.

ifMM :: Monad m => m Bool -> m a -> m (Maybe a)

If-statement for monads, with a monadic condition: if the condition holds, execute the given monad and return the result, otherwise return Nothing.

whileM :: Monad m => m Bool -> m () -> m ()

While-loop for monads. Any state passed between loop iterations should be carried by the monad.

transformWhile :: Monad m => a -> (a -> m (a, Bool)) -> m a

While-loop for monads with explicit state passing.

doMaybe :: Applicative m => Maybe (m ()) -> m ()

Execute the given Monad, or return () if not present.

Encoding state in IO monads

statefulIO :: (q -> i -> q) -> (q -> i -> o) -> q -> IO (i -> IO o)

Produce an IO monad that behaves like the given Mealy machine, with separate transition and output functions.

statefulIO' :: (q -> i -> (q, o)) -> q -> IO (i -> IO o)

Produce an IO monad that behaves like the given Mealy machine, with a combined transition/output function.