Lattest.Util.ModelParsingUtils

Documentation

readAutFile :: FilePath -> IO ([String], [String], Set String, String, Maybe [(String, IOAct String String, String)])

Read an .aut file representing an LTS from the provided filepath, parse its content and return: - [String]: Input Alphabet - [String]: Output Alphabet - Set.Set String: Set of all LTS states - String: Initial state - Maybe [(String, IOAct String String, String)]: List of LTS transition tuples as (InitialState, Action, EndState) NOTE: In order to parse actions correctly, inputs and outputs must end in _i and _o respectively. The first line of the .aut file must follow the structure des (initState,nEdges,nStates).

dumpLTSdot :: (Ord s, Show s, Ord i, Show i, Ord o, Show o) => FilePath -> [(s, IOAct i o, s)] -> IO ()

Build a .dot file representation of LTS transitions and save it in the specified File Path