Lattest.Util.Utils

Description

An unsorted collection of utility functions.

Boolean operators

(|||) :: (a -> Bool) -> (a -> Bool) -> a -> Bool infixr 2

Disjunction lifted to functions.

(&&&) :: (a -> Bool) -> (a -> Bool) -> a -> Bool infixr 3

Conjunction lifted to functions.

(||^^) :: (Monad m1, Monad m2) => m1 (m2 Bool) -> m1 (m2 Bool) -> m1 (m2 Bool) infixr 3

Disjunction lifted to a doubly nested monad.

(&&^^) :: (Monad m1, Monad m2) => m1 (m2 Bool) -> m1 (m2 Bool) -> m1 (m2 Bool) infixr 3

Conjunction lifted to a doubly nested monad.

Random functions

flipCoin :: RandomGen g => g -> Double -> (Bool, g)

Flip a coin for which the chance of result True is p, where p is clamped to [0,1].

takeRandom :: RandomGen g => g -> [a] -> (a, g)

Take a random element from the list. Gives an error for the empty list, so the user should check that first.

Maybe

takeJusts :: Foldable f => f (Maybe t) -> [t]

Filter out Nothings from a Foldable of Maybes, and returns only the Justs.

Set

takeArbitrary :: Set a -> Maybe (a, Set a)

Remove an arbitrary element from the given set, and return both that element and the remaining set, or Nothing if the given set was empty.

List utils

removeDuplicates :: Ord a => [a] -> [a]

Remove duplicates from a list by sorting, grouping and taking only the first element of each group. Returns ordered and filtered list.