Lattest.Model.Alphabet

Description

This module contains the definitions and semantics of different forms of observable actions, and inputs used in testing experiments.

Translating between Actions and Inputs

class Refusable act => TestChoice i act

If an input type is an TestChoice to a type of observable actions, this means that

given such an input, it is possible to derive the corresponding observable action, or sequence of actions, and

an observable action may (but also may not) correspond to an input.

Minimal complete definition

actToChoice, choiceToActs

choiceToActs :: TestChoice i act => i -> [act]

Derive the sequence of observable actions that correspond to an input.

actToChoice :: TestChoice i act => act -> Maybe i

If an observable action corresponds to an input, then derive that input.

Action types

Inputs and Outputs

data IOAct i o

Observable actions that may be either inputs provided to a system, or outputs from that system.

Constructors

In i

Out o

isInput :: IOAct i o -> Bool

Is the given action an input?

isOutput :: IOAct i o -> Bool

Is the given action an output?

fromInput :: IOAct i o -> i

Partially defined function that unpacks an input.

fromOutput :: IOAct i o -> o

Partially defined function that unpacks an outputs.

Observable Quiescence

Observable actions that may be either inputs provided to a system, or outputs from that system, where the output may also be an artificial output that represents an observed timeout, quiescence. For the theoretical background on observing quiescence, see

Jan Tretmans, Model based testing with labelled transition systems (Formal Methods and Testing), 2008

data Suspended o

Add observation of quiescence to a type of observable actions.

Constructors

Quiescence

OutSusp o

δ :: IOAct i (Suspended o)

type IOSuspAct i o = IOAct i (Suspended o)

Add observation of quiescence to the observed inputs and outputs.

asSuspended :: IOAct i o -> IOSuspAct i o

Convert an input or output to a type containing quiescence.

fromSuspended :: IOSuspAct i o -> IOAct i o

Partially defined function that unpacks an input or output from a type with quiescence.

Input Failures

For a Refusable type of observable actions, it is possible to decide whether the action was accepted or refused by the system that exhibited that observed action. Refusal of an input represents the system being unable to process the given input. This concept is described as an 'input failure'. For a theoretical background on input failures, see

Ramon Janssen, Refinement and partiality for model-based testing (Doctoral dissertation), 2022, Chapter 3 and 4

class Refusable act

The class of observable actions for which it is possible to derive whether the actions are accepted or refused. For types where refusal does not make sense, the default implementation is that all actions are accepted.

isAccepted :: Refusable act => act -> Bool

Is the action accepted or refused?

type IFAct i o = IOAct (InputAttempt i) o

Observable actions that may be either inputs provided to a system, or outputs from that system, where the inputs may be refused.

newtype InputAttempt i

Constructors

InputAttempt (i, Bool)

asInputAttempt :: IOAct i o -> IFAct i o

Convert an input or output to a type containing input failures.

fromInputAttempt :: IFAct i o -> IOAct i o

Partially defined function that unpacks an input or output from a type with input failures.

Combined Input Refusals and Quiescences

type SuspendedIF i o = IOAct (InputAttempt i) (Suspended o)

Input failure with observed quiescence. See IOSuspAct and IFAct for details.

asSuspendedInputAttempt :: IOAct i o -> SuspendedIF i o

Convert an input or output to a type containing input failures and quiescence.

fromSuspendedInputAttempt :: SuspendedIF i o -> IOAct i o

Partially defined function that unpacks an input or output from a type with input failures and quiescence.

STS

data SymInteract i o

Constructors

SymInteract (Gate i o) [Variable]

type SymGuard = SymBool

data SymAssign

data GateValue i o

Constructors

GateValue (Gate i o) [Value]

data Value

Constructors

IntVal Integer

BoolVal Bool

data Gate i o

Constructors

InputGate i

OutputGate o

data Variable

Constructors

Variable Symbol Type

addTypedVar :: Variable -> Value -> Model -> Model

data Type

Constructors

IntType

BoolType

data SymExpr

Constructors

BoolExpr SymBool

IntExpr SymInteger

equalTyped :: Variable -> Value -> Bool

assignedExpr :: Variable -> SymAssign -> Maybe SymExpr

assignment :: [(Variable, SymExpr)] -> SymAssign

noAssignment :: SymAssign