Lattest.Model.Alphabet

Description

This module contains the definitions and semantics of different forms of observable actions, and inputs used in testing experiments.

Translating between Actions and Inputs

class Refusable act => InputCommand i act

If an input type is an InputCommand to a type of observable actions, this means that

given such an input, it is possible to derive the corresponding observable action, or sequence of actions, and

an observable action may (but also may not) correspond to an input.

Minimal complete definition

actToInputChoice, inputChoiceToActs

inputChoiceToActs :: InputCommand i act => i -> [act]

Derive the sequence of observable actions that correspond to an input.

actToInputChoice :: InputCommand i act => act -> Maybe i

If an observable action corresponds to an input, then derive that input.

Action types

Inputs and Outputs

data IOAct i o

Observable actions that may be either inputs provided to a system, or outputs from that system.

Constructors

In i

Out o

isInput :: IOAct i o -> Bool

Is the given action an input?

isOutput :: IOAct i o -> Bool

Is the given action an output?

fromInput :: IOAct i o -> i

Partially defined function that unpacks an input.

fromOutput :: IOAct i o -> o

Partially defined function that unpacks an outputs.

Observable Timeouts

Observable actions that may be either inputs provided to a system, or outputs from that system, where the output may also be an artificial output that represents an observed timeout. For the theoretical background on observing timeouts, see the notion of quiescence in

Jan Tretmans, Model based testing with labelled transition systems (Formal Methods and Testing), 2008

data Timeout o

Add observation of timeouts to a type of observable actions.

Constructors

Timeout

TimeoutOut o

δ :: IOAct i (Timeout o)

type TimeoutIO i o = IOAct i (Timeout o)

Add observation of timeouts to the observed inputs and outputs.

asTimeout :: IOAct i o -> TimeoutIO i o

Convert an input or output to a type containing timeouts.

fromTimeout :: TimeoutIO i o -> IOAct i o

Partially defined function that unpacks an input or output from a type with timeouts.

Input Failures

For a Refusable type of observable actions, it is possible to decide whether the action was accepted or refused by the system that exhibited that observed action. Refusal of an input represents the system being unable to process the given input. This concept is described as an 'input failure'. For a theoretical background on input failures, see

Ramon Janssen, Refinement and partiality for model-based testing (Doctoral dissertation), 2022, Chapter 3 and 4

class Refusable act

The class of observable actions for which it is possible to derive whether the actions are accepted or refused. For types where refusal does not make sense, the default implementation is that all actions are accepted.

isAccepted :: Refusable act => act -> Bool

Is the action accepted or refused?

type IFAct i o = IOAct (Attempt i) o

Observable actions that may be either inputs provided to a system, or outputs from that system, where the inputs may be refused.

newtype Attempt i

Constructors

Attempt (i, Bool)

asInputAttempt :: IOAct i o -> IFAct i o

Convert an input or output to a type containing input failures.

fromInputAttempt :: IFAct i o -> IOAct i o

Partially defined function that unpacks an input or output from a type with input failures.

Combined Input Refusals and Timeouts

type TimeoutIF i o = IOAct (Attempt i) (Timeout o)

Input failure with observed timeouts. See TimeoutIO and IFAct for details.

asTimeoutInputAttempt :: IOAct i o -> TimeoutIF i o

Convert an input or output to a type containing input failures and timeouts.

fromTimeoutInputAttempt :: TimeoutIF i o -> IOAct i o

Partially defined function that unpacks an input or output from a type with input failures and timeouts.

STS

data SymInteract i o

Constructors

SymInteract (Gate i o) [Variable]

type SymGuard = SymBool

type SymAssign = Map Variable SymExpr

data GateValue i o

Constructors

GateValue (Gate i o) [Value]

data Value

Constructors

IntVal Integer

BoolVal Bool

data Gate i o

Constructors

InputGate i

OutputGate o

data Variable

Constructors

Variable Symbol Type

addTypedVar :: Variable -> Value -> Model -> Model

data Type

Constructors

IntType

BoolType

data SymExpr

Constructors

BoolExpr SymBool

IntExpr SymInteger

equalTyped :: Variable -> Value -> Bool