Lattest.Model.Automaton

Description

This module contains the definitions and semantics of automata models.

Syntactical Automaton Model

Definition

data AutSyn m loc t tloc

Syntactical automaton model, with locations and transitions. This is analogous to an automaton drawn on paper with points and arrows. Transitions are mapped to state configurations, see Lattest.Model.StateConfiguration. Furthermore, transitions contain transition labels, both on the outside and in the inside of the state configuration.

These labels are abstract and may be interpreted in various ways, e.g. a simple automaton model may directly have observable actions as labels, whereas a more complex automaton model may have symbolic data variables with guards, assignments, clocks for timing, etc.

locConf :: AutSyn m loc t tloc -> m loc

transRel :: AutSyn m loc t tloc -> loc -> Map t (m (tloc, loc))

Constructing Syntactical Automata

automaton :: (PermissionConfiguration m, Observable t, Ord t, Foldable fld) => m loc -> fld t -> (loc -> Map t (m (tloc, loc))) -> AutSyn m loc t tloc

Construct an automaton from an initial state configuration and a transition mapping

automaton' :: (PermissionApplicative m, Observable t, Ord t) => loc -> Set t -> (loc -> Map t (m (tloc, loc))) -> AutSyn m loc t tloc

Construct an automaton from an initial state and a transition mapping

Semantical Automaton Model

Definition

data AutSem m loc q t tloc act

Semantical automaton model. This model contains a reference to the syntactical automaton on which it is based, but it also contains a state configuration with semantical states.

The difference between the locations from the syntactical model and the states from the semantical model depends on the automaton semantics. E.g. for a simple automaton model, states and locations may be the same, whereas a more complex automaton model may have states consisting of syntactical locations combined with valuations for data variables, clocks for timing, etc.

stateConf :: AutSem m loc q t tloc act -> m q

syntacticAutomaton :: AutSem m loc q t tloc act -> AutSyn m loc t tloc

Constructing Syntactical Automata

semantics :: AutomatonSemantics m loc q t tloc act => AutSyn m loc t tloc -> (loc -> q) -> AutSem m loc q t tloc act

Infer a semantical model from a syntactical model, based on the appropriate instance of the AutomatonSemantics typeclass. Note that an automaton may be interpreted in multiple ways, so the type checker may need a hint on which semantical automaton is requested. This can be avoided by calling more specific, pre-typed variants of semantics in Lattest.Adapter.StandardAdapters.

Type Classes for Semantics

class Observable act

The Observable typeclass defines which types can be used as labels on transitions.

Minimal complete definition

implicitDestination

implicitDestination :: (Observable act, PermissionConfiguration m) => act -> m any

Defines the implicit state configuration reached by a given transition label if that label is omitted from a syntactical automaton. E.g. if a state contains no outgoing transition for an output label, that label is often considered to map to the forbidden state configuration.

class (Ord t, Observable act) => TransitionSemantics t act

TransitionSemantics expresses that the semantics of a syntactic transition can be expressed in terms of actions. E.g. symbolic transitions with interaction variables that can be expressed in terms of concrete observed values.

Minimal complete definition

asTransition

class StateSemantics loc q

StateSemantics expresses that the semantics of a syntactic location can be expressed in terms of state q. E.g. a location symbolic variables can be given in terms of valuations of these variables.

Minimal complete definition

asLoc

class StateConfiguration m => AutomatonSemantics m loc q t tloc act

Automaton semantics expresses that we can take steps, to move from one state configuration to another.

Minimal complete definition

after

after :: AutomatonSemantics m loc q t tloc act => AutSem m loc q t tloc act -> act -> AutSem m loc q t tloc act

Take a transition for the given action.

afters :: AutomatonSemantics m loc q t tloc act => AutSem m loc q t tloc act -> [act] -> AutSem m loc q t tloc act

Take a sequence of transitions for the given actions.

Finite Transition Labels

class TransitionSemantics t act => FiniteMenu t act

The class of automata with a finite list of transition labels on outgoing transitions for every state. This property is useful for e.g. test selection.

Minimal complete definition

asActions, locationActions

specifiedMenu :: (AutomatonSemantics m loc q t tloc act, Foldable m, Ord t) => FiniteMenu t act => AutSem m loc q t tloc act -> [act]

Menu of specified actions that are semantically present in the automaton.