Lattest.Model.Automaton

Description

This module contains the definitions and interpretations of automata models.

Syntactical Automaton Model

Definition

data AutSyntax (m :: Type -> Type) loc t tdest

Syntactical automaton model, with locations and transitions. This is analogous to an automaton drawn on paper with points and arrows. Transitions are mapped to state configurations, see Lattest.Model.BoundedMonad. Furthermore, transitions contain transition labels, both on the outside and in the inside of the state configuration.

These labels are abstract and may be interpreted in various ways, e.g. a simple automaton model may directly have observable actions as labels, whereas a more complex automaton model may have symbolic data variables with guards, assignments, clocks for timing, etc.

initConf :: AutSyntax m loc t tdest -> m loc

alphabet :: AutSyntax m loc t tdest -> Set t

trans :: Ord t => AutSyntax m loc t tdest -> loc -> t -> m (tdest, loc)

The transition relation as a function. Note that this function is partial, and only defined for transition labels in the alphabet of the automaton.

transRel :: AutSyntax m loc t tdest -> loc -> Map t (m (tdest, loc))

Constructing Syntactical Automata

automaton :: (BoundedConfiguration m, Completable t, Ord t, Foldable fld) => m loc -> fld t -> (loc -> Map t (m (tdest, loc))) -> AutSyntax m loc t tdest

Construct an automaton from an initial state configuration and a transition mapping

automaton' :: (BoundedApplicative m, Completable t, Ord t) => loc -> Set t -> (loc -> Map t (m (tdest, loc))) -> AutSyntax m loc t tdest

Construct an automaton from an initial state and a transition mapping

Semantical Automaton Model

Definition

data AutIntrpr (m :: Type -> Type) loc q t tdest (act :: k)

Semantical automaton model. This model contains a reference to the syntactical automaton on which it is based, but it also contains a state configuration with semantical states.

The difference between the locations from the syntactical model and the states from the semantical model depends on the automaton interpret. E.g. for a simple automaton model, states and locations may be the same, whereas a more complex automaton model may have states consisting of syntactical locations combined with valuations for data variables, clocks for timing, etc.

stateConf :: AutIntrpr m loc q t tdest act -> m q

syntacticAutomaton :: AutIntrpr m loc q t tdest act -> AutSyntax m loc t tdest

Constructing Syntactical Automata

interpret :: forall (m :: Type -> Type) loc q t tdest act. AutomatonSemantics m loc q t tdest act => AutSyntax m loc t tdest -> (loc -> q) -> AutIntrpr m loc q t tdest act

Infer a semantical model from a syntactical model, based on the appropriate instance of the AutomatonSemantics typeclass. Note that an automaton may be interpreted in multiple ways, so the type checker may need a hint on which semantical automaton is requested. This can be avoided by calling more specific, pre-typed variants of interpret in Lattest.Adapter.StandardAdapters.

Type Classes for Semantics

class Completable act

The Completable typeclass defines which types can be used as labels on transitions.

Minimal complete definition

implicitDestination

implicitDestination :: forall {k} m (any :: k). (Completable act, BoundedConfiguration m) => act -> m any

Defines the implicit state configuration reached by a given transition label if that label is omitted from a syntactical automaton. E.g. if a state contains no outgoing transition for an output label, that label is often considered to map to the forbidden state configuration.

class (Ord t, Completable act) => TransitionSemantics t act

TransitionSemantics expresses that the interpret of a syntactic transition can be expressed in terms of actions. E.g. symbolic transitions with interaction variables that can be expressed in terms of concrete observed values.

Minimal complete definition

asTransition

class StateSemantics loc q

StateSemantics expresses that the interpret of a syntactic location can be expressed in terms of state q. E.g. a location symbolic variables can be given in terms of valuations of these variables.

Minimal complete definition

asLoc

class BoundedMonad m => AutomatonSemantics (m :: Type -> Type) loc q t tdest act

Automaton interpret expresses that we can take steps, to move from one state configuration to another.

Minimal complete definition

after

after :: AutomatonSemantics m loc q t tdest act => AutIntrpr m loc q t tdest act -> act -> AutIntrpr m loc q t tdest act

Take a transition for the given action.

afters :: forall (m :: Type -> Type) loc q t tdest act. AutomatonSemantics m loc q t tdest act => AutIntrpr m loc q t tdest act -> [act] -> AutIntrpr m loc q t tdest act

Take a sequence of transitions for the given actions.

Finite Transition Labels

class TransitionSemantics t act => FiniteMenu t act

The class of automata with a finite list of transition labels on outgoing transitions for every state. This property is useful for e.g. test selection.

Minimal complete definition

asActions, locationActions

specifiedMenu :: forall (m :: Type -> Type) loc q t tdest act. (AutomatonSemantics m loc q t tdest act, Foldable m, Ord t, FiniteMenu t act) => AutIntrpr m loc q t tdest act -> [act]

Menu of specified actions that are semantically present in the automaton.

STS State data types

data IntrpState a

Constructors

IntrpState a Valuation

type Valuation = Map Variable Value

data STStdest

stsTLoc :: SymGuard -> SymAssign -> STStdest

Auxiliary Automaton Functions

reachable :: forall loc (m :: Type -> Type) t tdest. (Ord loc, Foldable m) => AutSyntax m loc t tdest -> Set loc

Compute the set of locations that is syntactically reachable from the initial location configuration. See reachableFrom.

reachableFrom :: forall loc (m :: Type -> Type) f t tdest. (Ord loc, Foldable m, Foldable f) => AutSyntax m loc t tdest -> f loc -> Set loc

Compute the set of locations that is syntactically reachable from the given locations.

Note that this not involve any interpretation of the automaton, e.g. if a location of symbolic automaton is only reachable via a transition with a guard that is always False, then that location is still considered to be reachable, even if a symbolic interpretation of that automaton can never reach that location for any trace of concrete values.

prettyPrint :: forall (m :: Type -> Type) tdest loc t. (Show (m (tdest, loc)), Show (m loc), Show loc, Show t, Ord loc, Foldable m) => AutSyntax m loc t tdest -> String

prettyPrintFrom :: forall (m :: Type -> Type) tdest loc t f. (Show (m (tdest, loc)), Show (m loc), Show loc, Show t, Ord loc, Foldable m, Foldable f) => AutSyntax m loc t tdest -> f loc -> String

prettyPrintIntrp :: forall {k} (m :: Type -> Type) tdest loc q t (act :: k). (Show (m (tdest, loc)), Show (m loc), Show loc, Show (m q), Show t, Ord loc, Foldable m) => AutIntrpr m loc q t tdest act -> String