Lattest.Model.StandardAutomata

Description

This module contains some simple automata types, and auxiliary functions for constructing them in a convenient manner. Note that the types are no more than type aliases, so if the automata type with your favorite combination of type parameters is not listed, don't hesitate to construct an automaton type (AutSyn or AutSem) yourself instead of using the StandardAutomata. The base function automaton for constructing automata is re-exported here for convenience.

Automaton construction

automaton :: (PermissionConfiguration m, Observable t, Ord t, Foldable fld) => m loc -> fld t -> (loc -> Map t (m (tloc, loc))) -> AutSyn m loc t tloc

Construct an automaton from an initial state configuration and a transition mapping

Alphabets

Auxiliary functions useful for constructing alphabets, intended for creating automata via automaton.

ioAlphabet :: (Traversable t, Ord i, Ord o) => t i -> t o -> Set (IOAct i o)

construct an alphabet of input-output-actions (IOAct) from separate alphabets of inputs and outputs

Transition relations

Auxiliary functions useful for constructing transition relations, intended for creating automata via automaton. ** Deterministic Transition Relations

detConcTransFromRel :: (Observable t, Ord loc, Ord t) => [(loc, t, loc)] -> Maybe (loc -> Map t (DetState ((), loc)))

Create a deterministic concrete transition relation from an explicit list of tuples, with the destination of transitions expressed as explicit states. Having multiple occurrences of a transition label is forbidden, i.e., leads to a Nothing result.

detConcTransFromMRel :: (Observable t, Ord loc, Ord t) => [(loc, t, DetState loc)] -> Maybe (loc -> Map t (DetState ((), loc)))

Create a deterministic concrete transition relation from an explicit list of tuples, with the destination of transitions expressed as deterministic state configuration. Having multiple occurrences of a transition label is forbidden, i.e., leads to a Nothing result.

detConcTransFromMaybeRel :: (Observable t, Ord loc, Ord t) => [(loc, t, Maybe loc)] -> Maybe (loc -> Map t (DetState ((), loc)))

Create a deterministic concrete transition relation from an explicit list of tuples, with the destination of transitions expressed as Just deterministic states, where Nothing is mapped to either forbidden or underspecified, depending on the transition label. Having multiple occurrences of a transition label is forbidden, i.e., leads to a Nothing result.

Non-Deterministic Transition Relations

nonDetConcTransFromRel :: (Observable t, Ord loc, Ord t, Applicative m, JoinSemiLattice (m ((), loc))) => [(loc, t, loc)] -> loc -> Map t (m ((), loc))

Create a non-deterministic concrete transition relation from an explicit list of tuples, with the destination of transitions expressed as explicit states. The state configuration must support non-determinism, and having multiple occurrences of a transition label is interpreted as non-deterministic choice between the destinations.

nonDetConcTransFromMRel :: (Observable t, Ord loc, Ord t, Applicative m, JoinSemiLattice (m ((), loc))) => [(loc, t, m loc)] -> loc -> Map t (m ((), loc))

Create a non-deterministic symbolic transition relation from an explicit list of tuples, with the destination of transitions expressed as explicit states. The state configuration must support non-determinism, and having multiple occurrences of a transition label is interpreted as non-deterministic choice between the destinations.

Create a concrete transition relation from an explicit list of tuples, with the destination of transitions expressed as non-deterministic state configuration. The state configuration must support non-determinism, and having multiple occurrences of a transition label is interpreted as non-deterministic choice between the destinations.

nonDetConcTransFromListRel :: (Observable t, Ord loc, Ord t) => [(loc, t, [loc])] -> loc -> Map t (NonDetState ((), loc))

Create a non-deterministic concrete transition relation from an explicit list of tuples, with the destination of transitions expressed as lists of locations, where the empty list is mapped to either forbidden or underspecified, depending on the transition label. Having multiple occurrences of a transition label is interpreted as non-deterministic choice between the destinations.

Alternating state configurations

Re-exports, so that test scripts don't need to import StateConfiguration separately

data FDL a

Free distributive lattice, or a positive boolean formula, i.e., a boolean formula with conjunctions and disjunctions over atomic propositions. The two elements top and bot can be interpreted as true and false.

atom :: a -> FDL a

A single state embedded in a free distributive lattice.

top :: FDL a

The free distributive lattice element ⊤, or true.

bot :: FDL a

The free distributive lattice element ⊥, or false.

(\/) :: FDL a -> FDL a -> FDL a

Disjunction on free distributive lattices.

(/\) :: FDL a -> FDL a -> FDL a

Conjunction on free distributive lattices.

Transition Functions

transFromFunc :: (Foldable fld, Ord t) => (loc -> t -> m (tloc, loc)) -> fld t -> loc -> Map t (m (tloc, loc))

Create a transition relation from a transition function. Warning: to use the resulting transition relation in an automaton, the function must be defined for all reachable states, and for all transition labels in the alphabet of the automaton.

concTransFromFunc :: (Foldable fld, Functor m, Ord t) => (loc -> t -> m loc) -> fld t -> loc -> Map t (m ((), loc))

Create a concrete transition relation from a transition function. Warning: to use the resulting transition relation in an automaton, the function must be defined for all reachable states, and for all transition labels in the alphabet of the automaton.

Automaton Semantics

Auxiliary functions for creating semantical automata AutSem. Note that most of these functions are no more than calls to semantics, instantiating the type of the semantical interpretation.

type ConcreteAutSem (m :: Type -> Type) q act = AutSem m q q act () act

Semantics of automata in which syntactical states and actions are directly interpreted as literal, semantical states and actions.

semanticsConcrete :: forall (m :: Type -> Type) t loc. (StateConfiguration m, Ord t, Show t, Show loc) => AutSyn m loc t () -> ConcreteAutSem m loc t

Interpret syntactical states and actions directly as literal, semantical states and actions.

type ConcreteTimeoutAutSem (m :: Type -> Type) q i o = AutSem m q q (IOAct i o) () (TimeoutIO i o)

Semantics of automata in which syntactical states and actions are directly interpreted as literal, semantical states and actions, but with timeouts as possible output observations.

semanticsQuiescentConcrete :: forall (m :: Type -> Type) i o loc. (StateConfiguration m, Ord i, Ord o, Show i, Show o, Show loc) => AutSyn m loc (IOAct i o) () -> ConcreteTimeoutAutSem m loc i o

Interpret syntactical states and actions are directly as literal, semantical states and actions, but with timeouts as possible output observations.

type ConcreteTimeoutInputAttemptAutSem (m :: Type -> Type) q i o = AutSem m q q (IOAct i o) () (TimeoutIF i o)

Semantics of automata in which syntactical states and actions are directly interpreted as literal, semantical states and actions, but with timeouts and input failures as possible observations.

semanticsQuiescentInputAttemptConcrete :: forall (m :: Type -> Type) i o loc. (StateConfiguration m, Ord i, Ord o, Show i, Show o, Show loc) => AutSyn m loc (IOAct i o) () -> ConcreteTimeoutInputAttemptAutSem m loc i o

Interpret syntactical states and actions are directly as literal, semantical states and actions, but with timeouts and input failures as possible observations.

semanticsSTS :: forall i o loc (m :: Type -> Type). (Ord i, Ord o, Ord loc, Show loc, Show i, Show o, Show (m (IntrpState loc)), StateConfiguration m, Show (m ((SymGuard, SymAssign), loc))) => STS m loc i o -> (loc -> IntrpState loc) -> AutSem m loc (IntrpState loc) (SymInteract i o) (SymGuard, SymAssign) (GateValue i o)

type STSIntrp (m :: Type -> Type) loc i o = AutSem m loc (IntrpState loc) (SymInteract i o) (SymGuard, SymAssign) (GateValue i o)