Lattest.Model.Symbolic.Internal.Sum

Sum types

type FreeSum a = FreeMonoidX (SumTerm a)

FreeSum represents a symbolic sum of terms of the type parameter a. The same term can occur multiple times.

newtype SumTerm a

Terms of a free-monoids of the form:

a0 <> a1 <> ... <> an-1

where <> will be interpreted as the arithmetic sum of terms:

a0 + a1 + ... + an-1

Constructors

SumTerm

Fields

summand :: a

Sum of Term and Sums

add :: Ord a => a -> FreeSum a -> FreeSum a

O(log n). Add a term to a sum.

subtract :: Ord a => a -> FreeSum a -> FreeSum a

O(log n). Subtract a term from a sum.

sum :: Ord a => FreeSum a -> FreeSum a -> FreeSum a

O(n+m). The sum of two sums.

sums :: Ord a => [FreeSum a] -> FreeSum a

The sum of a list of sums.

Multiply

multiply :: Ord a => Integer -> FreeSum a -> FreeSum a

O(n). Multiply the constant with the sum.